All the files are the ones that I changed.
Ability data allows you to assign different abilities to the ships that they can use. From there, when creating the turn-based mechanic and utilizing the grid. The range is something I wasn't sure how to implement, but I made instructions that most abilities are are few grid spaces ahead for the range they can fire. The main player abilities should be move, fire gun, missile, missiles, barrage onslaught, and, if you want at the end a railgun.  The order in the menu should be how I previously presented it. Enemies should get randomly generated abilities, or we can assign them for each level.